requirements.txt
langchain==0.0.340
chromadb==0.4.18
lark==1.1.8
openai==0.28.1
tiktoken
PyPDF2==3.0.1
streamlit==1.28.2
python-dotenv
pandas
python-docx==1.1.0